requirements.txt
transformers
torch
gradio
langchain
ibm_watson_machine_learning
huggingface-hub
ffmpeg
google-cloud-storage
google-cloud-aiplatform